pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "rdmo-dev"
authors = [
    {name = "Jochen Klar", email = "mail@jochenklar.de"},
]
requires-python = ">=3.8"
license = {file = "LICENSE"}
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Framework :: Django :: 2.2',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Operating System :: OS Independent',
    "Programming Language :: Python :: 3 :: Only",
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]
dependencies = [
    "django-debug-toolbar>=4.2.0"
]
dynamic = ["version"]

[tool.setuptools]
packages = ["rdmo_dev"]

[tool.setuptools.dynamic]
version = {attr = "rdmo_dev.__version__"}

[tool.ruff]
target-version = "py38"
line-length = 120
select = [
  "B",    # flake8-bugbear
  "C4",   # flake8-comprehensions
  "E",    # pycodestyle
  "F",    # pyflakes
  "I",    # isort
  "PGH",  # pygrep-hooks
  "RUF",  # ruff
  "UP",   # pyupgrade
  "W",    # pycodestyle
  "YTT",  # flake8-2020
]
ignore = [
  "B006",   # mutable-argument-default
  "B007",   # unused-loop-control-variable
  "B018",   # useless-expression
  "RUF012", # mutable-class-default
]

[tool.ruff.isort]
section-order = [
  "future",
  "standard-library",
  "pytest",
  "django",
  "rest_framework",
  "third-party",
  "rdmo",
  "first-party",
  "local-folder"
]

[tool.ruff.isort.sections]
pytest = ["pytest"]
django = ["django"]
rest_framework = ["rest_framework"]
rdmo = ["rdmo"]
